# Inflate into slice rather than Vec