@197g 2025-06-05
Also as the benchmarks show, there is still something funky going on with the flush allocation either way. -40.817% gains on Transparency? That seems frankly absurd from just that avoided allocation. So maybe we still have a misconception about the amount of data being shuffled somewhere in the decoding process.