@fintelia 2025-06-05
The improvement in the Transparency benchmark is caused by passing `self.data_stream[..self.filled + self.growth_bytes]` rather than just `self.data_stream`. If I revert that line but keep the rest, Transparency is unchanged from the contents of the main branch.

I believe the reason this helps is because after the buffer has been grown to 128+ KB, it is still better to alternate 8KB uncompress and 8KB unfilter, rather than operating on 128 KB chunks as the current code does.